busted: made "bg-like" colors be distinct from bg
this is a bit buggy at the moment - there are some situations where a
color config can fail to allow for colors that are "different enough"
from the background, yet also giving the proper contrast-ratio threshold
for all the other colors. i need to add a way for addBgColor() to bail
the whole scheme generation process and force it to start over.