fix threshold of "minimum contrast from CR midpoint"
i wanted to get something like "half the distance from the requested CR
to one", but the fact that "any contrast ratio" is represented as zero
threw me off initially. this will make "any contrast ratio" require a CR
of 1 from the midpoint, but that's effectively the same as zero, since
CR only goes down to 1.